Improved minor issue BLDC simulator: Hall sensors now convert to double outside
Improved a minor issue in the BLDC simulator. Now the hall sensor measurements
are converted from logical to double outside of the 'hall sensors' block, and
the BLDC hall output is now always a logical signal. This was done so that the
back propagation convert doesnt become an issue when using the 'encoder (HALL)'
block.